+++ MacroTracker.Models/FoodListItem.cs
@@ -16,11 +16,11 @@ namespace MacroTracker.Models
         public string FoodName { get; set; }
   
         public int Calories { get; set; }
-   
-        public int Proteins { get; set; }
-    
-        public int Fats { get; set; }
-       
-        public int Carbs { get; set; }
+
+        public decimal Proteins { get; set; }
+
+        public decimal Fats { get; set; }
+
+        public decimal Carbs { get; set; }
     }
 }
